# ENH: scipy.sparse.lil: tocsr accelerated