This PR accelerates the `tocsr` method of the `lil_matrix` in `scipy.sparse` by about 35%. This also accelerates other methods of the `lil_matrix` which use `tocsr`.