@jor- 2019-10-14
I can not run the benchmarks due to a compiler error in scipy/cluster/_hierarchy.c. However I did a simple benchmark by myself:
```
A = scipy.sparse.random(10**4, 10**4, density=0.5, format='lil')
%timeit A.tocsr()
```
results in 
```
9.01 s ± 1.1 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
for the old implementation and
```
5.9 s ± 324 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
for the new one.