@perimosocordiae 2019-10-14
I was able to run `python3 runtests.py --bench -s sparse` on a clean master branch without errors. Maybe something is up with you local repo?