@perimosocordiae 2019-10-16
I ran the benchmarks:
```
python3 runtests.py -t sparse.Conversion \
  --bench-compare=a04161d328bb4adf7b427ef52f00857daa656e1d \
  --bench-compare=7d9022349787cd4e0faef55a30cab22e91e37246 
```

with results:
```
      before           after         ratio
     [a04161d3]       [7d902234]
     <master~3>       <master~2>
+      7.14±0.3ms       18.3±0.1ms     2.56  sparse.Conversion.time_conversion('lil', 'csr')
+     7.31±0.07ms       17.9±0.4ms     2.45  sparse.Conversion.time_conversion('lil', 'csc')
+      7.54±0.3ms       18.3±0.2ms     2.42  sparse.Conversion.time_conversion('lil', 'coo')
+      7.97±0.2ms       18.1±0.2ms     2.27  sparse.Conversion.time_conversion('lil', 'bsr')
+      10.9±0.1ms       21.5±0.3ms     1.98  sparse.Conversion.time_conversion('lil', 'dia')
+      26.3±0.2ms       37.1±0.5ms     1.41  sparse.Conversion.time_conversion('lil', 'dok')
+         241±2μs         253±10μs     1.05  sparse.Conversion.time_conversion('csc', 'coo')

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
```

This particular benchmark is using a "poisson2d" sparse matrix generator function, which produces a 10000x10000 sparse matrix with 49600 stored elements (density ~= 0.0005) with roughly 3 nonzero elements per row. That means this PR's change will be looping 10k times, with two 3-element slice assignments each iteration.

For more random sparsity patterns we do indeed see a speedup with this PR, but it does seem that we have potential regression here. Maybe it's possible to select the appropriate construction method with a heuristic based on density?